Rebuild targets when any source dependency changes
Makefile now uses gcc flag -MM to generate a dependency list for
each source file to be included in the makefile. These dependency
files list the build target along with header file dependencies
in the same format expected in a Makefile target, therefore Makefile
can include the file directly.

However, the dependency file does not take directory structure into
account, therefore a sed command is required to modify the target
name to include the directory that it resides in.