Save and restore LR when entering and returning from an interrupt.
An interrupt would cause the current executing function to return
early on return from the interrupt, because lr was not set correctly
during return. ELR_hyp is set to the lr to be used after the return
from the interrupt, while lr is set to pc at the time of interrupt.
Therefore, lr should be saved on entry to the interrupt and restored
on return.

Fixes #1